refactor : 디렉터리 수정
